prueba5

+++ src/main/java/com/ipn/mx/modelo/entidades/Producto.java
@@ -6,7 +6,6 @@
 package com.ipn.mx.modelo.entidades;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -23,7 +22,7 @@ public class Producto implements Serializable{
     private int idProducto;
     private String nombreProducto;
     private String descripcionProducto;
-    private BigDecimal precio;
+    private float precio;
     private int existencia;
     private int stockMinimo;    
     private int idCategoria;
